requirements.txt
comtypes==1.1.7
tqdm==4.51.0
pick==1.0.0
termcolor==1.1.0
xlrd==1.2.0
mysql_connector_repackaged==0.3.1
